Declare eval_last_constant_dwarf_sub_expr with [u]int64_t not [s]size_t.
On 32bit platforms [s]size_t isn't the same as [u]int64_t causing:

abg-dwarf-reader.cc:358:1: warning: ‘bool abigail::dwarf_reader::eval_last_constant_dwarf_sub_expr(Dwarf_Op*, size_t, ssize_t&, bool&)’ declared ‘static’ but never defined [-Wunused-function]

	* src/abg-dwarf-reader.cc (eval_last_constant_dwarf_sub_expr):
	Declare expr_len as uint64_t and value as int64_t.

+++ src/abg-dwarf-reader.cc
@@ -356,8 +356,8 @@ die_is_at_class_scope(const read_context& ctxt,
 		      Dwarf_Die& class_scope_die);
 static bool
 eval_last_constant_dwarf_sub_expr(Dwarf_Op*	expr,
-				  size_t	expr_len,
-				  ssize_t&	value,
+				  uint64_t	expr_len,
+				  int64_t&	value,
 				  bool&	is_tls_address);
 
 static translation_unit::language
